requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-grappelli>=2.0,<3.0
djangorestframework>=3.0,<4.0
python-dateutil>=2.0,<3.0